Lab 2 Version 2.1

+++ src/kvsrv/common.go
@@ -8,7 +8,7 @@ type PutAppendArgs struct {
 	// Field names must start with capital letters,
 	// otherwise RPC will break.
 	Seqno int64
-	Cltno int64
+	Cltno int
 }
 
 type PutAppendReply struct {
@@ -18,7 +18,7 @@ type PutAppendReply struct {
 type GetArgs struct {
 	Key string
 	// You'll have to add definitions here.
-	Cltno int64
+	Cltno int
 }
 
 type GetReply struct {
